+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 # Change Log
 
-## Unreleased
+## Unreleased `0.13.0`
 
 - **Added**
-  - `#152`: As scp user, I want `scp.download` to return statistics so that i can further use the info.
+  - `#152`: As scp user, I want `scp.download` to return statistics so that I can further use the info.
+  - `#5`: As scp user, I want to call custom function for each downloaded file.
 - **Code Refactor**
   - `#107`: As dev, I want to refactor `console.warn`, `console.info`, `console.log` to `lib/core-output`.
   - `#154`: As dev, I want to change all `console.log` to `console.info` so that i can standardize the codes.